pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mail-processor"
version = "0.1.0"
description = "Non-interactive IMAP mail processor"
requires-python = ">=3.9"

dependencies = [
  "python-dotenv>=1.0"
]

[project.scripts]
mail-processor = "mail_pipeline.main:main"
[tool.setuptools.packages.find]
where = ["src"]